Split up useFilteredConsultants to smaller hooks
