+++ frontend/src/components/AvailabilityFilter.tsx
@@ -1,10 +1,10 @@
 "use client";
+import { useAvailabilityFilter } from "@/hooks/staffing/useAvailabilityFilter";
 import FilterButton from "./FilterButton";
-import { useFilteredConsultants } from "@/hooks/useFilteredConsultants";
 
 export default function AvailabilityFilter() {
   const { availabilityFilterOn, toggleAvailabilityFilter } =
-    useFilteredConsultants();
+    useAvailabilityFilter();
 
   return (
     <div className="flex flex-col gap-2">
